chore: split code in file
